+++ error.vue
@@ -1,58 +1,48 @@
 <script lang="ts" setup>
 import type { NuxtError } from '#app'
-import { Button } from '@/components/ui/button'
 
 const { error = { message: 'No error encountered' } } = defineProps({
   error: Object as () => NuxtError,
 })
 
-const l = useLoggerStore()
-
-function refresh() {
-  clearError({ redirect: '/' })
-}
-
-function clearAndRefresh() {
-  const auth = useAuthenticationStore()
-  const user = useUserStore()
-
-  auth._header.authorization = null
-  auth.loggedIn = false
-
-  user.user = null
-
-  refresh()
-}
-
 onMounted(() => {
-  // throw error
+  document.body.style.backgroundColor = '#180909'
 })
 </script>
 
 <template>
   <div
-    class="w-full h-[100vh] overflow-hidden justify-center items-center flex flex-col"
+    class="h-[100vh] w-[75%] max-w-[850px] overflow-hidden m-auto justify-center items-start flex flex-col"
   >
-    <h1 class="font-bold text-4xl text-red-500 text-center">
-      An error occurred<br>Oh noes :c
-    </h1>
-    <p class="font-semibold text-lg mt-4">
-      {{ error.message }}
-    </p>
-    <div class="flex flex-col gap-2 font-semibold mt-8">
-      <Button
-        size="lg"
-        @click="refresh"
+    <div class="inline-flex flex-row items-end w-full justify-between gap-4 flex-wrap ">
+      <div>
+        <h1 class="md:text-6xl text-4xl font-bold text-red-500">
+          {{
+            /* @ts-expect-error Generic Error */
+            error.statusCode
+          }}
+        </h1>
+        <h2 class="md:text-4xl text-2xl font-bold text-red-500">
+          {{ error.message }}
+        </h2>
+      </div>
+
+      <img
+        src="/shadfin_app_concept.svg"
+        class="size-24"
+        alt="Shadfin Logo"
       >
-        Refresh Page
-      </Button>
-      <Button
-        variant="destructive"
-        size="lg"
-        @click="clearAndRefresh"
-      >
-        Clear User Data & Refresh
-      </Button>
     </div>
+
+    <p class="mt-6 text-lg text-orange-400 font-semibold">
+      {{
+        /* @ts-expect-error Generic Error */
+        error.data
+      }}
+    </p>
+    <p
+      class="text-lg mt-4 text-orange-400 font-light font-mono"
+      v-html="error.stack"
+    />
   </div>
 </template>