+++ global.d.ts
@@ -1,7 +1,9 @@
-declare module "*.jpg";
-declare module "*.png";
-declare module "*.jpeg";
-declare module "*.gif";
-declare module "*.ass";
-declare module "*.md";
-declare module "*.txt";
+import './types/index'
+
+declare module '*.jpg';
+declare module '*.png';
+declare module '*.jpeg';
+declare module '*.gif';
+declare module '*.ass';
+declare module '*.md';
+declare module '*.txt';